Introduce measurements of accessing native data structure elements
The HTML form was changed to group benchmarks by type so that the functional interfaces
can be called consistently within a group of related measures:

  data_access: repeatedly retrieve elements of a data structure
  operator:    execute specific arithmetic operations
  iterate:     execute a function call or loop on a data structure
  process:     spawn, kill, link, monitor processes
  message:     measure send and receive message speeds
  data_manip:  transform a composite data structure (e.g., lists:sort)

Only the interfaces for data_access are provided with this check in (messaging and
function calls are not exposed via the HTML form, and the other categories do not
have code yet).

The main gen_server dk_bench_server now handles the following calls:

  mq_raw(NumMsgs) : send / receive tests
  mq_raw(NumMsgs, TimesToRun)

  exec(Fun, LoopCount) : execute a function repeatedly
  exec(Fun, LoopCount, TimesToRun)

  access(Fun, DataSize, LoopCount) : access a data structure repeatedly
  access(Fun, DataSize, LoopCount, TimesToRun)

The fun_calls module was extended to support the following functions for
repeatedly executing a function:

  function_call(LoopCount)  : Loop a native function decrementing LoopCount
  mfa_call(M, F, LoopCount) : Loop M:F(A) decrementing LoopCount

The fun_calls module was extended to support the following functions for
repeatedly accessing a data structure:

  list_nth    : Access lists:nth
  list_head   : Access [H|T] to get head
  binary_raw  : Access << _Skip:Pos*8, C, _Rest/tail >>
  binary_at   : Access binary:at(Pos)
  tuple_inx   : Access element(Pos, Tuple)

The data access calls allocate a randomly generated DataSize lengthed structure,
then access it Count times using a random position where possible.

Results are returned as before in a property list and are displayed as an HTML
bar graph representing amortized nanoseconds per operation.